feat: continue to extract git related stuff out of editor.rs
